Update gem release documentation
I cannot tell if we still use the prior documented approach, using a shared sul-devops-team service account, and I can't dig up any documentation of same on Consul or the googles. (Not to mention credentials.) Creating this PR to spark a discussion of how we want to handle this going forward.