amend dependency update deployment instructions
per discussion in slack, it's hard to get a true baseline where the only difference is the dependency updates (because we don't have a pre-update tag, and because non-`main` test branches may have been deployed to stage throughout the prior week).  so amend instructions suggest running one post-deployment test and discussing possible regressions in chat.

also, a bit of advice on running the test suite and when to consider a test genuinely failed.